+++ app/usecase/user_auth_repository.go
@@ -1,0 +1,9 @@
+package usecase
+
+type UserAuthRepository interface {
+	IssueToken(string, string) (string, error)
+	RegisterToken(string, string) (int, error)
+	GetPermissionName(string) (string, error)
+	ValidateToken(string) (bool, error)
+	RevokeToken(string) (int, error)
+}
